add: bag logic

+++ admin/src/hooks/useDeleteCategory.jsx
@@ -1,4 +1,5 @@
 import { useState } from 'react'
+import { useAuthContext } from './useAuthContext'
 import { useCategoriesContext } from './useCategoriesContext'
 
 export const useDeleteCategory = () => {
@@ -7,24 +8,24 @@ export const useDeleteCategory = () => {
     const [isLoading, setIsLoading] = useState(null)
 
     const { dispatch } = useCategoriesContext()
-    // const { user } = useAuthContext()
+    const { user } = useAuthContext()
 
     const deleteCategory = async (id) => {
         setIsLoading(true)
         setError(null)
         setSuccess(null)
 
-        // if (!user) {
-        //     setError('You must be logged in!')
-        //     return
-        // }
+        if (!user) {
+            setError('You must be logged in!')
+            return
+        }
 
         const response = await fetch(`${import.meta.env.VITE_API_URL}/api/category/${id}`, {
             method: 'DELETE',
-            // headers: {
-                // 'Content-Type': 'application/json',
-                // 'Authorization': `Bearer ${user.token}`
-            // }
+            headers: {
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
